rename redis_chash -> chash_redis
